# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "arc-storage",
        "version": "v1.0.2"
    },
    "host": "192.168.8.245:8081",
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/api/data/v1/history/arc": {
            "get": {
                "tags": [
                    "Arc Storage Data"
                ],
                "summary": "Return arc values. Querying from TDEngine",
                "operationId": "arc",
                "parameters": [
                    {
                        "name": "inside",
                        "in": "query",
                        "description": "inside swarm or not",
                        "required": false,
                        "type": "boolean",
                        "format": "boolean"
                    },
                    {
                        "name": "sensorids",
                        "in": "query",
                        "description": "多个ID逗号分隔",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "起始时间",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "终止时间",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "function",
                        "in": "query",
                        "description": "可选项,聚合查询",
                        "required": false,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "interval",
                        "in": "query",
                        "description": "可选项,聚合时间段的窗口",
                        "required": false,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "fill",
                        "in": "query",
                        "description": "可选项,数据填充格式",
                        "required": false,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\n\t\t- 可选项说明: SQL查询使用函数(聚合函数、选择函数、计算函数、按窗口切分聚合等)。\n\t\t- 不使用可选项，则输出查询到的所有数据。\n\t\t- function - 单个输出选择函数,推荐first, 参数:avg,sum,min,max,first,last。\n\t\t- interval - 聚合时间段的窗口,interval指定,最短时间间隔10毫秒(10a),推荐100ms。\n\t\t- fill     - 指定某一窗口区间数据缺失的情况下的填充模式,推荐使用PREV,详细查看文档。\n\t\t{\n\t\t\t\"code\": 0,\n\t\t\t\"msg\": \"OK\",\n\t\t\t\"data\": [\n\t\t\t\t{\n\t\t\t\t\t\"sensorid\": \"A00000000000\",\n\t\t\t\t\t\"data\": [\n\t\t\t\t\t\t{\n\t\t\t\t\t\t\t\"Time\": 1658209829000,\n\t\t\t\t\t\t\t\"arc\": []byte{0x94,0xC9,0x60,0x00,0xC2,0x48}]\n\t\t\t\t\t\t},\n\t\t\t\t\t\"count\": 1\n\t\t\t\t}\n\t\t\t]\n\t\t}\n\t\t",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer",
                                "format": "int32"
                            }
                        }
                    },
                    "400": {
                        "description": "\n\t\t{\n\t\t\t\"code\": 400,\n\t\t\t\"msg\": \"Bad Request\"\n\t\t}\t\t\n\t\t"
                    },
                    "404": {
                        "description": "\n\t\t{\n\t\t\t\"code\": 404,\n\t\t\t\"msg\": \"Not Found\"\n\t\t}\t\t\n\t\t"
                    },
                    "429": {
                        "description": "\n\t\t{\n\t\t\t\"code\": 429,\n\t\t\t\"msg\": \"Too Many Requests:\"+ id\n\t\t}\t\t\n\t\t"
                    },
                    "503": {
                        "description": "\n\t\t{\n\t\t\t\"code\":503,\n\t\t\t\"msg\":\"Request Timeout\"\n\t\t}\t\n\t\t"
                    }
                }
            }
        },
...
        "/api/data/v1/history/sensorids": {
            "get": {
                "tags": [
                    "Arc Storage Data"
                ],
                "summary": "Get information of sensor ids",
                "operationId": "sensorids",
                "parameters": [
                    {
                        "name": "inside",
                        "in": "query",
                        "description": "inside swarm or not",
                        "required": false,
                        "type": "boolean",
                        "format": "boolean"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\n\t\t{\n\t\t\t\"code\": 0,\n\t\t\t\"msg\": \"OK\",\n\t\t\t\"data\": [\n\t\t\t\t\"A00000000000\"\n\t\t\t]\n\t\t}\n\t\t",
                        "schema": {
                            "$ref": "#/definitions/SensorIDResponse"
                        }
                    },
                    "404": {
                        "description": "\n\t\t{\n\t\t\t\"code\": 404,\n\t\t\t\"msg\": \"Not Found\"\n\t\t}\t\t\n\t\t"
                    },
                    "429": {
                        "description": "\n\t\t{\n\t\t\t\"code\": 429,\n\t\t\t\"msg\": \"Too Many Requests:\"+ id\n\t\t}\t\t\n\t\t"
                    },
                    "503": {
                        "description": "\n\t\t{\n\t\t\t\"code\":503,\n\t\t\t\"msg\":\"Service Unavailable\"\n\t\t}\t\n\t\t"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "SensorIDResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "xml": {
                        "name": "Code"
                    },
                    "format": "int32"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "format": "string"
                    },
                    "xml": {
                        "name": "Data",
                        "wrapped": true
                    }
                },
                "msg": {
                    "type": "string",
                    "xml": {
                        "name": "Msg"
                    },
                    "format": "string"
                }
            },
            "xml": {
                "name": "SensorIDResponse"
            }
        },
...
    },
...
}